clean up
deleting all the scripts that don't really matter to me anymore that I wrote.
They were just for collecting  objdata from the .blend files and the 302.xls sheet. But they were supposed to interact with the Blender API and when I removed that they just stopped working- and I got sick of them because I already got the data I needed and so I'm just getting rid of them. If I want, I can rewrite them with the express purpose of extracting data-- but there are other more important things.